requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.13
requests
python-dotenv
google-generativeai
